Don't pass return values to hook callback.
If the promise chain resolves to a value and we pass it to the callback, it
will be treated as an error and we'll chase our tails for days trying to
figure out how we got a non-error error value. Just use a little lambda to
call the callback with no arguments.